refactor : pay model-domain 분리
