+++ blackWhiteDeliveryDriver/src/main/java/com/sparta/blackwhitedeliverydriver/model/pay/entity/PayStatusEnum.java
@@ -1,0 +1,5 @@
+package com.sparta.blackwhitedeliverydriver.model.pay.entity;
+
+public enum PayStatusEnum {
+    SUCCESS, REFUND;
+}
