+++ blackWhiteDeliveryDriver/src/main/java/com/sparta/blackwhitedeliverydriver/model/pay/repository/PayRepository.java
@@ -1,8 +1,8 @@
-package com.sparta.blackwhitedeliverydriver.pay.repository;
+package com.sparta.blackwhitedeliverydriver.model.pay.repository;
 
-import com.sparta.blackwhitedeliverydriver.order.entity.Order;
-import com.sparta.blackwhitedeliverydriver.pay.entity.Pay;
-import com.sparta.blackwhitedeliverydriver.user.entity.User;
+import com.sparta.blackwhitedeliverydriver.model.order.entity.Order;
+import com.sparta.blackwhitedeliverydriver.model.pay.entity.Pay;
+import com.sparta.blackwhitedeliverydriver.model.user.entity.User;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
