--- blackWhiteDeliveryDriver/src/main/java/com/sparta/blackwhitedeliverydriver/pay/entity/PayStatusEnum.java
@@ -1,5 +1,0 @@
-package com.sparta.blackwhitedeliverydriver.pay.entity;
-
-public enum PayStatusEnum {
-    SUCCESS, REFUND;
-}
