+++ blackWhiteDeliveryDriver/src/main/java/com/sparta/blackwhitedeliverydriver/pay/service/PayService.java
@@ -8,21 +8,21 @@ import com.sparta.blackwhitedeliverydriver.pay.dto.PayReadyResponseDto;
 import com.sparta.blackwhitedeliverydriver.pay.dto.PayRefundRequestDto;
 import com.sparta.blackwhitedeliverydriver.pay.dto.PayRefundResponseDto;
 import com.sparta.blackwhitedeliverydriver.pay.dto.PayRequestDto;
-import com.sparta.blackwhitedeliverydriver.order.entity.Order;
-import com.sparta.blackwhitedeliverydriver.order.entity.OrderProduct;
-import com.sparta.blackwhitedeliverydriver.order.entity.OrderStatusEnum;
-import com.sparta.blackwhitedeliverydriver.order.entity.OrderTypeEnum;
-import com.sparta.blackwhitedeliverydriver.pay.entity.Pay;
-import com.sparta.blackwhitedeliverydriver.pay.entity.PayStatusEnum;
-import com.sparta.blackwhitedeliverydriver.user.entity.User;
-import com.sparta.blackwhitedeliverydriver.user.entity.UserRoleEnum;
+import com.sparta.blackwhitedeliverydriver.model.order.entity.Order;
+import com.sparta.blackwhitedeliverydriver.model.order.entity.OrderProduct;
+import com.sparta.blackwhitedeliverydriver.model.order.entity.OrderStatusEnum;
+import com.sparta.blackwhitedeliverydriver.model.order.entity.OrderTypeEnum;
+import com.sparta.blackwhitedeliverydriver.model.pay.entity.Pay;
+import com.sparta.blackwhitedeliverydriver.model.pay.entity.PayStatusEnum;
+import com.sparta.blackwhitedeliverydriver.model.user.entity.User;
+import com.sparta.blackwhitedeliverydriver.model.user.entity.UserRoleEnum;
 import com.sparta.blackwhitedeliverydriver.user.exception.ExceptionMessage;
 import com.sparta.blackwhitedeliverydriver.order.exception.OrderExceptionMessage;
 import com.sparta.blackwhitedeliverydriver.pay.exception.PayExceptionMessage;
-import com.sparta.blackwhitedeliverydriver.order.repository.OrderProductRepository;
-import com.sparta.blackwhitedeliverydriver.order.repository.OrderRepository;
-import com.sparta.blackwhitedeliverydriver.pay.repository.PayRepository;
-import com.sparta.blackwhitedeliverydriver.user.repository.UserRepository;
+import com.sparta.blackwhitedeliverydriver.model.order.repository.OrderProductRepository;
+import com.sparta.blackwhitedeliverydriver.model.order.repository.OrderRepository;
+import com.sparta.blackwhitedeliverydriver.model.pay.repository.PayRepository;
+import com.sparta.blackwhitedeliverydriver.model.user.repository.UserRepository;
 import com.sparta.blackwhitedeliverydriver.pay.util.HttpUtil;
 import com.sparta.blackwhitedeliverydriver.pay.util.PayUtil;
 import java.time.Duration;
